MODIFY : 스마트파인더 컴포넌트 분리
