+++ src/pages/ProductList/SmartFinder.scss
@@ -1,0 +1,157 @@
+.product-list-page {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  .product-list-container {
+    width: 50%;
+    display: flex;
+    flex-direction: column;
+
+    .product-list-title {
+      color: gray;
+      margin-bottom: 62px;
+    }
+    .separate-bar {
+      border-bottom: 1px solid rgba(0, 0, 0, 0.3);
+    }
+
+    .smart-finder-container {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+
+      h2 {
+        color: gray;
+        margin: 10px;
+      }
+
+      .smart-finder {
+        border-radius: 5px;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        width: 100%;
+        padding: 30px 70px;
+        background-color: rgb(230, 230, 230);
+
+        .finder-price-container,
+        .finder-category-container,
+        .finder-color-container {
+          margin-bottom: 30px;
+
+          .finder-title {
+            color: gray;
+            font-size: 15px;
+          }
+        }
+
+        .finder-select-bar {
+          width: 150px;
+          height: 10px;
+          background-color: gray;
+        }
+
+        .price,
+        .category,
+        .color {
+          color: gray;
+          display: flex;
+          flex-direction: column;
+          align-items: center;
+          margin-right: 5px;
+        }
+
+        .price:hover,
+        .category:hover {
+          div {
+            background-color: red;
+          }
+          color: red;
+          cursor: pointer;
+        }
+
+        .focusStyle {
+          div {
+            background-color: red;
+          }
+          color: red;
+          cursor: pointer;
+        }
+
+        .finder-price {
+          display: flex;
+
+          input {
+            outline: none;
+          }
+        }
+
+        .finder-category {
+          display: flex;
+        }
+        .finder-color {
+          display: flex;
+
+          .finder-color-bar {
+            width: 70px;
+            height: 30px;
+            background-color: gray;
+            border-radius: 5px;
+          }
+
+          .black {
+            background-color: black;
+          }
+          .gray {
+            background-color: gray;
+          }
+          .red {
+            background-color: red;
+          }
+          .blue {
+            background-color: blue;
+          }
+          .green {
+            background-color: green;
+          }
+          .pink {
+            background-color: pink;
+          }
+          .orange {
+            background-color: orange;
+          }
+          .yellow {
+            background-color: yellow;
+          }
+          .beige {
+            background-color: beige;
+          }
+          .white {
+            background-color: white;
+          }
+
+          .color:hover {
+            div {
+              border: 3px solid red;
+            }
+            color: red;
+          }
+
+          .color-red:hover {
+            div {
+              border: 3px solid black;
+            }
+            color: red;
+          }
+        }
+
+        .search-result {
+          align-self: flex-start;
+          color: gray;
+        }
+      }
+    }
+  }
+}
