+++ src/pages/test/App.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function App() {
+  return <div></div>;
+}
+
+export default App;
